// angular/webapp/src/app/booking-history/model/Reservation.ts
export class Reservation {
    reservationId: number = 0;
    rentedByEmailid: string = "";
    reservationDate: string = "";
    reservationStartDate: string = "";
    reservationEndDate: string = "";
    carId: string = "";
    carImage: any=null;
    carName: string = "";
    carType: string = "";
}